reddit-scraper/requirements.txt
# Reddit API 包装器
praw==7.7.1

# 环境变量管理
python-dotenv==1.0.0

# 数据处理（可选，用于高级分析）
pandas==2.1.4

# 进度条显示（可选）
tqdm==4.66.1
